reduce re rendering of page within header or story prop changes
