+++ app/javascript/pages/country/stories/stories.js
@@ -1,5 +1,6 @@
 import { createElement } from 'react';
 import { connect } from 'react-redux';
+import { getAdminsSelected } from 'pages/country/utils/filters';
 
 import StoriesComponent from './stories-component';
 import actions from './stories-actions';
@@ -8,12 +9,21 @@ export { initialState } from './stories-reducers';
 export { default as reducers } from './stories-reducers';
 export { default as actions } from './stories-actions';
 
-const mapStateToProps = () => ({
-  totalAmount: 'Nan',
-  percentage: 'Nan',
-  startYear: 2011,
-  endYear: 2014
-});
+const mapStateToProps = state => {
+  const adminData = {
+    location: state.location.payload,
+    countries: state.countryData.countries,
+    regions: state.countryData.regions,
+    subRegions: state.countryData.subRegions
+  };
+  return {
+    totalAmount: 'Nan',
+    percentage: 'Nan',
+    startYear: 2011,
+    endYear: 2014,
+    locationNames: getAdminsSelected(adminData)
+  };
+};
 
 const StoriesContainer = props =>
   createElement(StoriesComponent, {
